add minimal core module
