+++ src/lib.rs
@@ -4,43 +4,10 @@
 
 // Sub-modules:
 
+/// Core functionality
+pub mod core;
+
 /// An abstraction over SQL engines
 pub mod sql;
 
 // Global constants and other lookups:
-
-// ## API Module
-
-#[derive(Debug)]
-pub enum RelatableError {
-    /// An error in the configuration of a ChangeSet:
-    ChangeError(String),
-    /// An error in the Relatable configuration:
-    ConfigError(String),
-    /// An error that occurred while reading or writing to a CSV/TSV:
-    // CsvError(csv::Error),
-    /// An error involving the data:
-    DataError(String),
-    /// An error generated by the underlying database:
-    // DatabaseError(sqlx::Error),
-    /// An error from an unsupported format
-    FormatError(String),
-    /// An error in the inputs to a function:
-    InputError(String),
-    /// An error that occurred while reading/writing to stdio:
-    IOError(std::io::Error),
-    /// An error that occurred while serialising or deserialising to/from JSON:
-    // SerdeJsonError(serde_json::Error),
-    /// An error that occurred while parsing a regex:
-    // RegexError(regex::Error),
-    /// An error that occurred because of a user's action
-    UserError(String),
-}
-
-impl std::fmt::Display for RelatableError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{:?}", self)
-    }
-}
-
-impl std::error::Error for RelatableError {}
